requirements.txt
fastapi
uvicorn
pydantic
streamlit
openai
python-dotenv